Update components for new data, issues #19 and #21
